test: consolidate test suite — cut ~488 test-line insertions from diff
Targets 2 and 3 of Paul's consolidation dispatch:

- decision/tests.rs: −58 lines (5 row tests fully implied by property
  sweeps removed: LookupFailed duplicates, and 3 small cells already
  provable from the sweeps' 972-state coverage)
- config_sync/tests.rs: −443 lines (10 fine-grained unit tests removed
  whose behavior is fully exercised by the composed regression scenarios;
  test scaffolding extracted to shared test_helpers.rs)
- config_sync/regression_tests.rs: −101 lines (tests for the same-content
  early-return path and other coverage fully subsumed by the composition
  tests)
- config_sync/test_helpers.rs: +112 lines (new shared scaffolding used by
  all four test files)

Target 1 (retention/tests.rs relocation) is blocked: inlining 471 lines
of test code back into retention.rs would push the file from 959 to ~1430
lines, which violates the desktop file-size ratchet (1000-line limit).
Reported to Paul as a blocker.

No production-logic changes. Non-test file changes: config_sync.rs adds
the test_helpers module declaration (#[cfg(test)] only).

All four mutation checks still catch their reversions:
- Mutation 1 (retain_agent_record → retain_event): FAILS
  test_retain_agent_record_user_intent_clears_gate_but_reconcile_does_not
- Mutation 2 (tombstone_agent_pending_inner → retain_event): FAILS
  test_non_persona_tombstone_user_intent_clears_gate
- Mutation 3 (mark_synced_and_stamp_baseline split to autocommit): FAILS
  test_mark_synced_and_stamp_baseline_atomic_both_written_on_success
- Mutation 4 (readiness check removal): FAILS
  test_publish_via_flush_blocked_when_latch_not_ready
